// ui/src/services/cradle/models/OptimizedEntryResponse.ts
// @ts-nocheck
/* tslint:disable */
/* eslint-disable */
/**
 * CRADLE
 * Threat Intelligence Knowledge Management
 *
 * The version of the OpenAPI document: 2.10.2-beta.9bd46310 (2.10.2-beta.9bd46310)
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface OptimizedEntryResponse
 */
export interface OptimizedEntryResponse {
    /**
     * 
     * @type {number}
     * @memberof OptimizedEntryResponse
     */
    readonly id?: number;
    /**
     * 
     * @type {string}
     * @memberof OptimizedEntryResponse
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof OptimizedEntryResponse
     */
    readonly type?: string;
    /**
     * 
     * @type {string}
     * @memberof OptimizedEntryResponse
     */
    readonly subtype?: string;
    /**
     * 
     * @type {string}
     * @memberof OptimizedEntryResponse
     */
    readonly color?: string;
}

/**
 * Check if a given object implements the OptimizedEntryResponse interface.
 */
export function instanceOfOptimizedEntryResponse(value: object): value is OptimizedEntryResponse {
    if (!('name' in value) || value['name'] === undefined) return false;
    return true;
}

export function OptimizedEntryResponseFromJSON(json: any): OptimizedEntryResponse {
    return OptimizedEntryResponseFromJSONTyped(json, false);
}

export function OptimizedEntryResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): OptimizedEntryResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'name': json['name'],
        'type': json['type'] == null ? undefined : json['type'],
        'subtype': json['subtype'] == null ? undefined : json['subtype'],
        'color': json['color'] == null ? undefined : json['color'],
    };
}

export function OptimizedEntryResponseToJSON(json: any): OptimizedEntryResponse {
    return OptimizedEntryResponseToJSONTyped(json, false);
}

export function OptimizedEntryResponseToJSONTyped(value?: Omit<OptimizedEntryResponse, 'id'|'type'|'subtype'|'color'> | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'name': value['name'],
    };
}